25.03.04
Solve BOJ 2Prob

+++ 알고리듬 풀이/백준/10. 기하) 직사각형과 삼각형/수학은 체육과목 입니다_15894.cpp
@@ -23,7 +23,7 @@ using namespace std;
 
 int main(void)
 {
-    int n;
+    long long int n;
     cin >> n;
 
     cout << 4 * n << endl;
